- removing dar-help which was not more useful since a while because of internationalization of messages
dar-help was written by Chris Martin, to simplify the output of -h options for
dar, dar_slave, dar_xform and dar_manager. With dar-help the output was formatted
to 80 columns leaving the update of the -h messages very simple to carry on.

However, since the use of gettext, dar-help cannot work as the breaking of lines
depends on the use of the locale. A great thank to Chris Martin for his help
simplifying dar's maintenance.