- starting implementation of the "skip backward when file changed" feature
at this step, the feature works without escape sequence marks (-at) option
and if slicing is used, only if the skip backward does not imply the change
of slice. This second point is OK, as changing of slice may cause problem
if the slice is already burt or moved away.