- fixed bug that could lead a file to be reported as different from the one in filesystem (-d option) if it had changed of size during backup and has been saved properly a subsecquent time (--retry-on-change option)
when libdar detects a file has changed and the "retry-on-change" feature is activated
it is saved a second time, but did not checked whether its size changed or not.
Then, during archive comparison (-d option), libdar could report a difference between the archive
and the filesystem because of the wrong size recorded in the archive, while the corresponding
file could be restored properly at the same state it has in the filesytem.