- continuing the "skip backward when file changed" feature implementation
now, skipping backward is possible over an escape class (escape sequences
used for sequential reading). Still and forever it will not be possible to
skip backward if slicing is used and the skip points to a previous slice.
Last, remains a problem to fix concerning the fact that a file may have its
size reduced... this is a bug that will be addressed in 2.4.x