speeding up database listing
this patch adds a "read-only" status to the database constructor. This
state implies the already "partial" status, and avoid the ability to
dump() a database opened this way.
In partial mode, the database header is loaded, and the remaining is
stored in memory (not creating the corresponding objects hierarchy),
in order to be able to dump it as is if only the header has to be modified
(like changing the archive name, location, or option to pass to dar for
examples). In read-only mode, only the header is read, the rest is even
not read which saves CPU cycles as no decompression is required for that
part.