- fixing possible cause of bug SF #3547658.
The cache layer adapts its size to its ability to effectively be useful.
When comes the time to increase the cache size, upon failure of allocating a
large enough block of memory, the cache object throws an exception and
is no more able to handle subsequent writes which may occurs when upper
layer data is flushed in response to that exception. In that case, the
write request on this cache object tries to copy data to the now NULL buffer
internally used for caching.

The fix here, is to set back cache size to the previous value when it fails to
allocate a larger buffer, record that this value is the maximum to not exceed not
to waste ressource trying increase over that limi and be able to have write()
call working in degraded more when the cache object definitively fails to find
a memory block available for work, and in consequence has thrown an exception
to report the problem.