fixing the source of dar_manager database corruption
corruption occurs when deleting the first archive of a base that
only contains a file as "removed". This "removed" information is
propagated to the next archive which keep alive this entry in the base
while the parent directory is no more referred by any archive if it
also only exists in the just removed archive. This is the existence in
the base of that parent directory not part of any archive that triggers
the bug detection condition.