fixed an out of range access in a backward compatible routine that is never used due other better mechanism that has been implemented later.
This routine tends to take care of old archive transformed by
dar_xform released starting 2.4.0 and thus having a recent slice
header but still having an old archive layout with archive table only
at the beginning of the first slice (archive header). Since 2.4.x,
this table is also present at end of the last slice (archive trailer)
and a recent slice header let dar understand that the archive is recent
and can be opened by the end. But doing that the archive trailer starts
by the "droot" string which is what a catalogue starts by probably
meaning we have to read the archive header at the beginnning of first
slice because the archive is not that recent (archive format < 8).

Well, this mechanism is not perfect and has been superseeded by the
ability for sar class (which manages slicing) to keep generating old slice
header upon request. When dar_xform reads an archive which format is
older than "8" it produce an archive asking sar to generate an old
slice header (format "7"). This avoids the situation described previously
where the slice header is recent while the archive layout is not.

Long explanation for reminder about a fix of one byte that has absolutely
no impact on dar/dar_xform/dar_slave functionning!