fixing typo in full-from-diff target that lead to a syntaxical error
the quote '\' of the bang character (!) lead the command to fail on syntax
error. It is effectively only necessary whenusing that string on command line
from a tcsh shell (not from a bash shell). It has no reason to be present here
in included file, which type of file is directly read by dar not through any
shell.