adding debugging features to trigger sha1 generation problem reported bug
test_hash_fichier class get a new method to avoid writing down the data
and only compute the hash (sha1 or md5) in order to compare the result
with sha1sum. A script has been also built to understand the pattern when
(file size, etc.) that lead to such difference between ligcrypt and
sha1sum. Bug still under investigation because not stable scenario found
to reproduce it so far.