fixing bug met while doing a verbose listing in sequential read
In verbose mode, dar show the archive size. But to determine the
archive size libdar in certain circumpstances skip to the end of
the archive get the offset and returns at the original offset. When
reading in sequential read mode skipping through a pipe (stdin
for example) to eof implies dropping all data in cache and lead to
that part of the archive (that was in cache) to be unaccessible
which lead libdar to report data corruption.