adding reading by block for non ftp transfer to preserve existing connection
when dar needs to skip to another position, it set the callback to abort libcurl
which as the drawback to close the control connection, implying reauthentication,
this can hit some rate limiting or maximum session per user limits. To avoid that
and at the cost of network performance, a reading per block is used as an
alternative for sftp protocol, in that case when skipping, libdar only waits the
end of the block transmission for libcurl to return and then skip to another place.
Unfortunately, this does not avoid control session disconnection for FTP which
thus keep using the unlimited reading with libcurl