fixed bug: hard linked inode melted to another one when merging the result of a previous merging operation
hard linked inode are indexed by a so called "etiquette" which is an integer number. All
directory entry pointing to the same inode have the same etiquette. Etiquette start at zero
and are incremented by one at each new hard linked inode found.

When merging two archives, both have potentially the same etiquette numbers, to avoid mixing
then, the second archive read has all its etiquettes number increased by the number of
hard linked inode seen in the first archive. In other word, if 3 hard linked inode are found
in the first archive, they are numbered 0, 1 and 2. If 2 hard linked inode are found in the
auxilliary archive (where they are numbered 0 and 1) they get renumbered as 3 and 4 in the
resulting catalogue.

Then, comes the overwriting policy that may decide to not keep hard linked inode 1 if is
overwritten by another hard linked inode from the second archive (by etiquette 3, for example).
The resulting archive will have etiquette 0, 2, 3 and 4. so far so good.

When this resulting archive is in turn used as the first archive of the merging operation,
its etiquette get added as is into the resulting catalogue. The auxilliary archive of reference
get its etiquette shifted by the number of etiquette found in the first archive but due to
the missing 1 the etiquette 0 of the second archive is stored as etiquette 4 which already
exists in the resulting catalogue. Where from the possible merge of different hard linked
inodes that may have absolutely nothing related.

The fix brought here is to normalize the resulting archive, that's to say remove the hole
in the etiquette numbering (in the example above this means renumber the etiquette 4 as 1)
before applying the overwriting policy. The resulting archive may have holes in the etiquette
numbering that's not will not be anymore a problem if used a archive of reference for a
subsequent merging operation thanks to this bug fix.

Doing that way also fixes already existing archive that have non continuous etiquette
numbers so they can be used as source for subsequent merging without problem using libdar
having this bug fixed.