fixing bug that lead dar improperly generating an archive when aborted by user
the problem was a wrong offset of the beginning of the internal catalogue recorded
in the archive.

This wrong offset was due to the fact escape layer is lying over a cache layer. When
user interruption is caught during the filesystem read, no problem occured. The
bug expressed only when the interruption was caught while writing data to archive:

When the cache layer received a block of data to write down from the above escape
layer, it may end filling its internal buffer while not all data provided by the
escape layer fit in it. Thus the cache layer flushes the data to its below layer
(sar/trivial_sar which manages slicing). At this point the exception could be
triggered.

In that situation, the exception led the cache to abort keeping the buffer intact
containing some of the data passed from the escape layer, which receiving in turn
the exception considered none of the data passed to the below cache layer to be
written (because the call failed due to exception). However in 2.5.x for speed optimization
in regard to multi-threading tentative (poor performance gain, and disabled by default)
the cache layer computes what should be the offset of the underlying layer, leading
the cache layer offset being different from what the escape layer thinks it has.

When comes the time to get the offset where the catalogue will be dropped the return
value by escape layer was taken from what the cache should have but not from what it
really had. This wrong value was encoded in the archive terminator where from the
impossibility to read the catalogue at that wrong place on direct access mode.

The fix consist in modifying the cache layer to ignore all data block passed for writing
when this operation lead to the raising of an exception. This way what the any other
layer relying on cache layer may think/compute or record of the offset of the cache layer
will always stay coherent to the read offset cache layer has.