fixing bug in the filtering mecanism based on a list of file
for efficient search the list of file is sorted by a modified
lexicographical ordering function. While the lookup relies
on a binary search using the default lexicographical ordering function
leading in some situation to not find a file while it is present in the
list.

The consequence is that when the list provide files to include
some where not included and when the list was provided to list
a file to exclude some were not excluded as they ought to be