fixed bug in file listing (-[ option) leading some directories and their content to be excluded in a somehow rare condition
CONTEXT:
the provided list is sorted following a modified sort order that puts
at first the '/' character in order to have a directory appearing just before the
possible item located in it.

ordering the list addresses two needs:
- lookup efficiency (binary search)
- including a directory if it is not listed but some items of the list should
  take place in that directory

The binary such uses two variables "min" and "max", which are indexes on the
sorted list. The search stops when max - min <= 1
at that time item number max is compared with the path to consider, then
against the item number min. If non matches, we look at the max item to see
if it is located in the path to consider. If not, the entry can be
skipped (and all its content if it is a directory) else the path to consider
is a directory which is not present in the list *but* for which some of its
sub-file/sub-directory are listed: so the operation (backup for example) must
recurse into it.

THE BUG:
the problem arise when all items of the list are greater (according the
modified sort order) than the path to consider. In that case the binary search
stops with min == 0 and max == 1. In that situation the test explained above
that checks for list entry included in that directory is broken. The test
should be done against the item pointed to by min, not by max.

THE FIX:
at the end of the binary search and before the subdirectory check, if min is
still at zero and if the path to consider is less than the item pointed to by
min, max is set to min (to zero) in order to have the next step
(subdirectory test) to succeed in that condition too.