adding a method to cat_mirage that disable hard link reduction to normal inode
in repair mode, a the first hard link to a given inode is used for saving before
any other hard link to the same inode to be read. Thus at the time of the saving
the hard link count of the inode is 1 which means the inode can be saved as is,
that's to say without the hardlink overhead. This is true in if the archive has
been completely read, which is not the case in repair mode (sequentially reading
the damaged archive), this method allow the repair operation to flag each first
born hard link toward a given inode to always dump the inode as a hard linked
inode, future hard link to that inode will then have data to point to at archive
reading time.