adding a warning in repair mode when EA and FSA could not be read and copied properly
in that situation, the EA or FSA is removed from the repaired archive for the concerned
file, after a warning about file corruption has been reported. The removed EA and FSA
will be stored in the catalogue at the end but the same information along the archive
for sequential reading has already been written and refers to EA and FSA that have been
removed. The add message states in case this particular situation occurs, that the
concerned file will generate EA/FSA error in sequential read mode only