fixing bug about negative dates stored as huge positive numbers in old archive format
dar 2.5.12 and below in 2.5.x branch could not read archive
generated by dar 2.4.x and below (unless in infinint compilation mode)
when the old archive included a file which date(s) was returned by the
system as a negative integer at the time of the backup. Note that if
dar can now read old archive in that particular case, such date stay
recorded in the dar archive as very far in the future and not in the
past, because 2.4.x and below blindly assumed the system would always
return a positive integer as number of second since 1970. Since 2.5.12
release, when the system provides a negative date the date is assumed
as zero (Jan 1970) with user agreement.