fixing bug when writing to file when no space is left available
the write() system calls fails when the requested amount of data to write
is larger than the available space. It was assumed the write() call would
first write as much data as possible then at next call return a failure,
which is not the observed behavior.

To be more robust on the way the write() system call may return in case of
lack of space, dar_split now divides by two the amounts of data to write at
once when the system calls returns that there is not room to write the data.
It does so up to the time the amount of data to write is 1 byte and write()
returns an error. Only at that time it is assumed all data could be wrote
and that user shall be asked to do something to provide space to write data
to.