renaming user_interaction::warning() in message() to support backward compatibility to API v5
in API v5 user_interaction provides a warning_with_more feature which is implemented
in the warning() call, inherited_warning() being provided to user for the warning implementation.

as libdar5::user_interaction inherits from libdar::user_interaction it would not be possible
to keep the same method name from libdar for libdar5::user_interaction intercepts calls to
libdar::warning() to provide as it the ability to pause() every N lines. For that reason
renaming warning() in message() and inherited_warning() as inherited_message() in libdar
let the libdar5::user_interaction implement the warnng_with_more feature in inherited_message
and provide an inherited_warning as in API v5

the warning_with_more feature will be present in shell_interaction in libdar starting version 6.x.x