fixed wrong storage size stored in generated archives
once a file has been written to archive, the compression
layer flushing took place after the reading of the current
offset in the archive, used to calculate the file storage size.
This led some data pending in memory in the compression layer
that were written afterward and a wrong value for storage size.

storage size value is only used when non compression is used
to limit the amount of data to read from the archive when
restoring a file (For compressed files, the compressed data
contains a mark which when all compressed data have been passed
to the decompression engine emulate the end of file, thus the
storage size info is not necessary nor used in that case).