fixed man page about sparse-file handling while merging
it was written that setting --sparse-file-min-size to zero in addition to -ah
would let dar store sparse file of the archive of reference as normal file in
the resulting archive. But setting it to zero disables the sparse_file convertion
(same as if -ah was not provided). To achieve this goal one need to provide
-ah and give to --sparse-file-min-size a value larger than all file sizes of
the archive. 1E is a good value for dar built with 64 bits, but 1Z or 1Y or larger
could also be used, though those large integer do not fit in 64 bit integer and
will lead dar to ask using a infinint-based build of dar/libdar