fixing bug met when listing an archive using the same object that was used to create it
detecting whether the archive is isolated is not possible in that case
as it implies having access to the low layer level (contextual) of the
archive, which are not available just after archive creation (they are
available when reading an existing archive using a brand-new object).

Is this case, the archive is never isolated (it has just been created)
but it is not "exploitable" at the opposit of archive read using a brand
new object. So we use this status to find the "isolated" status instead
of trying to access the archive low layers which would fail anyway