avoid issuing warning about EA to be overwriting in some corner case
at restoration time, when an file on filesystem has not EA at all
the default 'EA merge' operation is replaced by the EA overwrite
operation for efficiency. Though, this action leads to issuing a
warning (unless -w option is set) to confirm the overwriting by the
user which is useless in that particular case as there is nothing to
overwrite.

The fix here is to avoid issuing this warning in that particular case