fixing bug in the routine removing files for local filesystem,
routine used at archive creation time to remove an existing archive (after user
confirmation), or at restoration time used to remove file that had been
removed since the archive of reference was done. The file to remove was
always removed from the current directory (missing the path part), most
of the time this was leading to the error message "Error removing file
 ...: Not such file or directory". I could also lead to incorrectly
removing files (not directory) located in the directory from which
dar was run.