adding signature block size information to class cat_delta_signature, increasing archive format version
it is necessary to know which block size has been used to create asignature
when comparing archive and filesystem. in some cirumstances, the comparison
leads to compare the stored signature with one calculated based on the file
data present in filesystem where from this information need to be saved in the
archive.
Trying to get this information from the librsync signature datastructure was
first risk (outside librsync API and subject to change in the future) and second
it required to build a generic_rsync sooner in the process to obtain this
information which would have been a bigger impact on libdar code