fixed bug: dar failed creating an archive on its standard output
dar failed with the error message "Skipping backward is not possible on a pipe"
the cache object layer did not validate that during the skippability
evaluation that even if the new position is contained in the cached data
some of this cached data may have already been passed (written) to the cached object.
if the offset is before the already written data, it will lead the
cached object to skip_backward which for a tuyau object (pipe) is not
possble, leading to the failure of the archive creation.
The skippability evaluation was wrong and should have return false
in that situation.

The fix now takes care of that condition and cache::skippability
when skipping backward is contained inside the cache and the position
is before the last written data returns the backward skippability of
the cached object.