workaround bug for delta patch generated by dar up to and including 2.6.7
the sparse_flag was not clear for file saved as delta patch while the
hole detection mecanism was not used to store the resulting delta patch into
the archive because it does not makes sense to do use it for such data.

at archive reading libdar seeing the flag was creating adding the sparse
file restoration layer that is looking for metadata inlined in the file
that is used to store hole information while there is not such metadata
to expect.

In some rare conditition, the delta patch may contain such escape sequence
that flag the beginning of inlined metadata. This will lead libdar to modify
the resulting data in a way that the CRC stored in the archive does no more
matches the read patch or the metadata is of a unknown type which in both case
lead the patch to be unreadable.

The fix is to ignore this sparse flag when the file data is a delta patch
which avoids adding a sparse_file layer when reading the patch