Since delta patch are no more flagged a sparse file reactivating the compression ratio display for delta patch
the condition that display the compression ratio of file did not include
delta binary patch. Due to the fix that made the sparse detection mechanism
never used for such type of saved data, the compression ratio was no more showed
for the delta binary patched files.

This fix is to activate back the display of compression ratio for delta patch