fixed bug in dar_split
when writing do multiple destination, a read block from source can be written
to its destination. This can fail either with an error (no byte written), or
fail with an amount of written bytes less than the amount to be written.
In this later case, dar_split takes into account the written bytes and move
to the remaing bytes to be written but still consider the amount of bytes to
be written has not change (instead if reducing the number by the bytes written
previously).
This fixe reduce this amount accordingly.
The impact in such condition could be an memory violation and crash of dar_split
or a corruption of the copied data.
When a tape of filesystem is full, Linux filesystem report an error and do no
copy any byte, so this situation under Linux is not likely to occur. However not
all system behave the same so the possibility of this bug to occur is not null