fixing bug in dar_manager part of libdar
when computing the list of archive to use to restore an entry, the
et_present status should only be considered when it is the initial
(or oldest) status available for an entry.
But when considering the next archive status, during sanity checks
it is wrongly assumed to be bug when the status of that file in the
previous archive is et_present. This effectively a bug condition
except if this is the initial status (the file has this status in the
oldest archive and we are considering the second archive for that file)

The fix takes into account this particular case and avoid throwing an
exception in that case, which let dar_manager wrongly report a bug
condition.