rewriting tronconneuse to receive a crypto_module as constructor argument
the pure virtual method that had to be implemented in crypto_sym have
been move to crypto_module class (new parent of crypto_sym).
the tronconneuse constructor now receive an unique_ptr on a crypto_module
object that implements those formely virtual method implemented in its
derived classes.

the objective is to first:
share the crypto stuff between tronconneuse and parallel_tronconneuse
and second it was necessary to be able to duplicate as much as there is
parallel workers those crypto_module to avoid having shared data beteen
workers (and performance penalty due to synchronization that would be
requested to access these shared data)