fixing bug that should affect current 2.6.x stable release
it is met when merging and changing compression algorithm. If the new
compression algorithm gives less good result than the original, the resaved_uncompress
feature is triggered and the file is read once again from the archive.

The fix consists to delaying the recording of the file position and
storage size in the new archive after the resaved_uncompressed has eventually taken
place, because else it modifies the place of the data in the source archive
which provides wrong data and fails due to CRC mismatch.