the salt is now computed by the crypto_sym we can reorder the archive creation sequence
the target is to be able to regenerate the crypto layer if libgcrypt reports
a weak key due to the pkcs5 hash and salt. As the salt is stored in the archive
header, we could not regenerate the hashed password using a new salt before.
Now thanks to this reordering we are able to validate the key strength

this will be the objective of the next commit