fixing bug met when restarting writing down file data due to poor compression or file changed during backup and previous file had EA without FSA
the problem has been identified with block_compressor class of future 2.7.0 branch
when compression is used and a file has to be resaved due to poor compression or
file has changed while it was read for backup *and* if the previous saved entry
had EA without FSA *and* no slice boundary is met *and* no hash computation is
requested *and* no encryption is asked (!)

Upon bad compression ratio or changed file during backup, dar tries to skip back
in the archive (which is not possible with compression or hash computation). To
be able to skip back, before saving the inode content the first time, dar records
the current offset, where eventually it will have to skip back upon such condition,
the code assumed that all data as synced (all compressed data written to the
compressed layer) which was wrong when the previous entry saved its EA (always
compressed), not followed by FSA (never compressed). In such situation the
offset is possibily wrong and if dar has the conditions to skip back it will
overwrite the EA of previous file (totally or partially).

the fix is to add a sync_write before reading the offset where to rewind.

The block_compressor layer had some more strict code control and triggered
this bug during non-regression tests of 2.7.0, the problem however also concerns
the compressor class of 2.6.0 and previous versions

This is a somehow very improbable bug, but not impossible to meet.