fixing missing reset of eof flag in block_compressor class
impacted lzo* compression algo in streaming and block compression as well as
any other compression algorithm in block compression when reading
backup in sequential mode. Dar signaled CRC error because no data could
be read from the compression layer the EOF not being reset. Direct access
mode jumps (skip() call) to the location of each file which clears the eof
flag.

block compression is used for multi-threading and for lzo in any mode