fixing bug in block_compressor met when uncompressing empty files
if --mincompr is set to zero, even empty files are flagged as being
compressed. When time comes to uncompress them, the block_compressor
canot find a block header which triggers an exception.

Change made detects end of file before trying to read an new block
header, which avoids triggering an exception and produces the expected
uncompressed result (an empty file).