fixing source of wrong offset data writing due to the cache layer after truncate() call
when truncate() is invoked on a cache object and the truncate position
is located inside the cache while some data is still pending writing
just before that position, the below layer has not yet reached
the offset requested for truncate. After having cleared/setup the
cache to match the truncate request, propagating unconditionnaly
the truncate to the layer below, leads it to skip forward by the
amoun of date still pending for writing. Data that will thus be
written later at the truncate position and not before it, leading
to corrupted data.

At testing time, the computed CRC does not match the stored CRC
and the error is reported.

The fix is to only propagate the truncate if no data is pending
in the cache layer for writing.