fixing bug in escape::truncate leading to self reporting bug escape.cpp line 858
the escape implementation was not updating its internal position
from the below layer once truncate was propagated which triggered
this bug as it was assumed a subsequent skip() call should succeed
to that position.

The fix is to get back to the truncate() expected behavior, and at
this occasion flush the escape class internal write_buffer cache when
necessary.