fix when using sequential read on a pipe and dar has to remove files at restoration time
when using differential backup, dar stores file that have been removed
since the previous backup. These can only be detected once all backup has
completed by comparing the created catalogue with the one of the backup
of reference.

This information is thus only stored at the end of the archive

when restoring data in sequential-read mode, dar read the archive content
using the metadata in-lined all along the archive. But once completed it
need to read the catalogue at the end of the backup to check if there is
any file to remove, this is done by reading setting up a simplified
catalogue in memory from the catalogue in archive by only taking into
account directory and "detruit" objects (the later being used to flag files
to be removed).

This simplified catalogue is then passed to the upper layer as if it
was the continuation of the in-lined metadata which leads the restoration process
to walk to a subset of the directory tree already restored just to "chdir" to
the place where files have to be removed.

If the directory are left intact with all their EA and FSA attributes,
this second restoration if such directory will need to fetch EA and/or FSA
for restoration, those may be very earlier in the archive leading dar to
skip backward.

When sequential-read is used on top of a anonymous pipe (passing the
backup to dar's stdin for example), dar aborts the restoration reporting
that failing backward on a pipe is not possible.

To avoid this, this fix drops away the EA and FSA of the directory
when creating this simplified catalogue. Those EA and FSA have already
been restored earlier so there is no reason to restore them twice and
this also avoids having dar trying to skip backward to read them.