adding early memory release feature for catalogue class
destroy directory content once read, or compared. Once the read
or compare operation has completed, the catalogue is unusable (and empty)
This feature is not compatible when read a catalogue as reference, nor
when performing several actions on a backup (on-fly-isolation for example)

escape_catalogue inherit from catalogue and let the possibility to sequentially
read a backup with very reduced memory requirement as the catalogue is built
and releases all along the backup process.