avoiding using std::make_shared with libthreadar
std::make_shared with clang 10.0.1/FreeBSD 12 does
not tolerate class having a destructor explicitely
able to throw exception. (libthreadar::barrier, for
some good reason). Removing the explicit declaration
from the constructor leads to invocation of terminate()
in case of exception when compiling with gcc, which is
not what we want. This we set the std::shared_ptr
by mean of its reset() method.

By the way, doing so makes sens here as we test the
result of the object allocation and throw an Ememory
exception in case of allocation failure.

The previous code, would not have launched a std::bad_alloc
exception which is not what we want to provide to the user
(all thrown exception should be child or inherited child of
class libdar::Egeneric)