fixing bug in when using hashing with remote repo
hash_fichier class missed flushing the main file (it did for
the hash file). For local filesystem there is nothing to flush at
libdar level, but when using sftp or ftp, the libcurl objects has to be
flushed. In certain circumstances not doing so lead the subthread called
where libcurl is run stuck on mutex because it had completed passing data
to libcurl and is waiting for new data. In that situation the subthread
cannot be stopped. calling terminate() on the fichier_libcurl() object
does that precisely and it was missing in hash_fichier