fixing bug about the patch base checksum used to validate we apply a patch to the correct file
since format 11.2 the patch base cheksum evaluation has been moved earlier in the
backup procecss of each file and this information has been moved from the cat_delta_signature
to the cat_file class which is one of the first metadata wrote to backup. Having
the patch base crc in cat_delta_signature, which is wrote to disk after the data/patch,
was breaking the restoration in sequential read mode, as before applying a patch (and
reading the data/patch bits) the patch base crc (located after) was not yet available
and validation that we are about to apply the patch to the correct file, failed.

However during this change of holding class and archive backup location, the base crc was
not properly setup when the reference data was already a patch. Instead of using the CRC
of the previously patched data (stored in the bakup of reference for validation after patching),
the value used was the CRC of the patch itslef as also stored in the backup, leading
dar to fail patching the file (even if it was in the expected data to apply the patch on)

to the use of already existing backup in format 11.2 even if this fix does not change
the archive format, we increment to version 11.3 and will add workaround code to bypass
this check when format is 11.2