fixing bug in thread_cancellation class
when destroying a thread_cancellation object on a thread which was requested
to be cancelled, the status was kept forever in the preborn list, which purpose
is not for that but to record cancellation for thread which no thread_cancellation
object was yet run.

just deleting thread_cancellation reference from the info list is better, the
cancellation information for that tid will be forgotten once all thread_cancellation
object for that thread will be deleted.