fixing bug met when trying to explode an undisclosed path
when trying to explode an undisclosed path ending with a single /
an exception was not thrown and the last / was just ignored, while was
not with two or more ending slashes.

To build such an undiscloded path the input must have an additional
ending / which is removed when creating the path object.

Thus
- user input syntax A/ was OK (creating a valid path "A")
- user syntax A// was also OK but lead to a modified path "A/" due to
  the lack of exception flagging the path as invalid
- and the syntax A/// (or more /) failed to be "exploded" leading
  the path to stay undisclosed as "A///"

At filter time evaluation, in the second case the provided mask was
built with "A//" based fs_root, while the files under evaluation
were build based on "A/" path, which for file exclusion (-P option)
led the match to not be seen and expected excluded files or directory
to not be excluded from the backup operation.

This fix solve the issue with the second case, but zooming out,
it does not seems a good idea to try exploding the fs_root as valid
path in libdar code, due to the fact this operation is not done
neither when building the masks in dar command-line tool. This will
be reviewed and API documented accordingly.