in early memory release, avoid releasing mirages and directories containing mirages
so far, the cat_mirage was deleted when exiting a directoryu in early memory release,
if a next reference to that mirage was found in another directory, the catalogue::corres
map was still referring the now deleted entry, this was a bug. In fact we must first of all
not remove such mirage entry to keep alive the pointed-to cat_etoile the next
cat_mirage will have to build upon.