fixing bug met when reeading a ciphered data in sequential mode with the help of an external catalog
data and EA located right after the elastic buffer cannot be accessed and are reported
as corrupted, due to a flush_read needed (really?) in the context of dar_slave (thus not
in sequential read mode).

This commit condition the flush_read to the presence of a zapette object in the stack where from
is read the data.