fixing bug met when restoring a gpg signed archive in sequential read mode
when the archive is signed, once all in-lined metadata has been loaded, libdar
loads the internal catalogue to fetch only the detruit objects and the directories
that lead to them, this operation also drops EA and FSA of directories to not try
to restore them again (which would moreover require to skip back in the archive
to fetch them, something which is not possible in sequential read mode). so far,
so good.

But, when the archive is signed the internal catalogue is fully loaded to compare
it with the embedded gpg signature (and eventually report an error). However in that
case, when it comes to restore detruit objects as the EA and FSA have not been removed
dar tries to restore them for each subdirectory leading to a detruit object, which
fails and report the error.

The fix is to add a method to catalgue class that drives to remove all EA and FSA of
the existing objects. This is in complement of the constructor described above that
only loads detruit objects (and drops EA and FSA of directories at the same time).