fixing bug met while reading archive summary with slice in sequential read mode
dar/libdar was asking for the last slice even if it was present, because of a
request to skip_to_eof() which in sequential-read mode on a sar objects cannot
detect the slice flag which tells if the slice is or not the last of the backup.

Worse, the catalogue was not loaded because it need an operation to grab the inlined
metadata of the archive to setup the catalogue in memory.

Adding the init_catalogue() had solved both problems, loading the catalogue in memory
and avoiding asking to skip at end of file in sequential read mode